add api erros and responsive design

+++ src/components/logout/Logout.vue
@@ -4,10 +4,13 @@
 <script>
 export default {
     created() {
-        
         if (this.$store.getters.loggedIn) {
-            this.$store.dispatch('destroyToken')
-            this.$router.push({name: 'login'})
+            if(confirm('Do you want to leave?')){
+                this.$store.dispatch('destroyToken')
+                this.$router.push({name: 'login'})
+            } else {
+                this.$router.push({name: 'tasks'})
+            }
         } else {
             this.$router.push({name: 'login'})
         }
